Watch Extension now perform test before enabling the button
This commit is about making sure that the webserver that actually opens the garage door is available.

This is how the test works:
A HTTP request against a test page is perform prior to enabling the button that actually opens the garage door. Upon HTTP response code 200 the button will be enabled otherwise the test gets repeated. The solves two problems:

1.) Now the user can see from the button state whether the action can be expected to work or fail.
2.) Before implementation of this test there often times there was significant delay between pressing the button and the door acually opening. That could have triggered the use to repeatedly press the button, which might lead to unexpected results.